test(core-models): add tests for the new F* specific core-models.
