+++ hax-lib/core-models/std/Cargo.toml
@@ -7,6 +7,9 @@ edition = "2024"
 hax-lib.workspace = true
 core-models = { path = "../core-models" }
 
+[dev-dependencies]
+proptest = "1.9.0"
+
 [lib]
 doctest = false
